requirements.txt
django==4.1.5
mysqlclient==2.1.1
pymysql==1.0.2
cryptography==39.0.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0